+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - Fixed bug where heros quickly became lvl 100 once they past lvl 28
 - Bash cooldown now applies to the two bash items
 - Arcane Echo now considers Nimbus an ultimate when setting cooldown
+- Anti Perma-Stun Mechanic is now a mutator in the advanced map
 
 ## 9.7
 - Added new heros and updated talent replacement system